+++ src/test/java/domain/LadderResultTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import common.exception.message.ExceptionMessage;
 import common.exception.model.ValidationException;
+import domain.ladder.LadderResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
