+++ src/test/java/domain/PlayerNamesTest.java
@@ -2,6 +2,8 @@ package domain;
 
 import common.exception.message.ExceptionMessage;
 import common.exception.model.ValidationException;
+import domain.player.PlayerName;
+import domain.player.PlayerNames;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
